requirements.txt
streamlit==1.47.0
google-generativeai>=0.3.1
odfpy>=1.4.1
Pillow>=9.4.0
pandas>=1.5.0
numpy>=1.24.0
PyYAML
bcrypt
requests>=2.28.0
mysql-connector-python==8.0.32